Move FatalError and FailFast to Contracts package
