+++ src/EditorFeatures/TestUtilities/Async/AsynchronousOperationBlocker.cs
@@ -6,7 +6,7 @@
 
 using System;
 using System.Threading;
-using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.ErrorReporting;
 
 namespace Roslyn.Test.Utilities;
