+++ src/ExpressionEvaluator/Core/Test/ResultProvider/Microsoft.CodeAnalysis.ResultProvider.Utilities.csproj
@@ -41,12 +41,6 @@
     <Compile Include="..\..\..\..\Compilers\Core\Portable\InternalUtilities\Debug.cs">
       <Link>Compiler\InternalUtilities\Debug.cs</Link>
     </Compile>
-    <Compile Include="..\..\..\..\Compilers\Core\Portable\InternalUtilities\FailFast.cs">
-      <Link>Compiler\InternalUtilities\FailFast.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\..\Compilers\Core\Portable\InternalUtilities\FatalError.cs">
-      <Link>Compiler\InternalUtilities\FatalError.cs</Link>
-    </Compile>
     <Compile Include="..\..\..\..\Compilers\Core\Portable\InternalUtilities\EnumField.cs">
       <Link>Compiler\InternalUtilities\EnumField.cs</Link>
     </Compile>
