+++ src/LanguageServer/Protocol/Handler/SourceGenerators/SourceGeneratorRefreshQueue.cs
@@ -6,6 +6,7 @@ using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.CodeAnalysis.Threading;
 using Roslyn.LanguageServer.Protocol;
